+++ src/unabto/unabto_connection.c
@@ -290,7 +290,7 @@ static size_t mk_connect_rsp(uint8_t* buf, uint8_t* end, uint16_t seq, uint32_t 
         ptr = insert_capabilities(ptr, end, 1 /*unenc*/);
     }
 
-    insert_length(buf, ptr - buf);
+    insert_length(buf, (uint16_t)(ptr - buf));
     return ptr - buf;
 }
 
@@ -376,7 +376,7 @@ static void send_rendezvous_socket(nabto_socket_t socket, nabto_connect* con, ui
     {
         size_t len = ptr - buf;
 
-        insert_length(buf, len);
+        insert_length(buf, (uint16_t)len);
 
         if (seq) {
             NABTO_LOG_DEBUG((PRInsi " RENDEZVOUS Send to " PRIep ": seq=%" PRIu16, MAKE_NSI_PRINTABLE(0, con->spnsi, 0), MAKE_EP_PRINTABLE(*dest), seq));
@@ -977,7 +977,7 @@ uint16_t unabto_count_active_connections()
 
 bool nabto_write_con(nabto_connect* con, uint8_t* buf, size_t len) {
     con->stats.packetsSent++;
-    con->stats.bytesSent += len;
+    con->stats.bytesSent += (uint32_t)len;
 #if NABTO_ENABLE_TCP_FALLBACK
     if (con->type == NCT_REMOTE_RELAY_MICRO) {
         unabto_tcp_fallback_error status = unabto_tcp_fallback_write(con, buf, len);
@@ -1113,7 +1113,7 @@ void send_connection_statistics(nabto_connect* con, uint8_t event) {
     }
 
     length = ptr - nabtoCommunicationBuffer;
-    insert_length(nabtoCommunicationBuffer, length);
+    insert_length(nabtoCommunicationBuffer, (uint16_t)length);
     send_to_basestation(nabtoCommunicationBuffer, length, &nmc.context.gsp);
 }
 
@@ -1149,7 +1149,7 @@ void send_connection_ended_statistics(nabto_connect* con) {
     }
 
     length = ptr - nabtoCommunicationBuffer;
-    insert_length(nabtoCommunicationBuffer, length);
+    insert_length(nabtoCommunicationBuffer, (uint16_t)length);
 
     send_to_basestation(nabtoCommunicationBuffer, length, &nmc.context.gsp);
 }
